Fix: 결제 정보 없을 때 추가
